+++ src/frontend/globalState/models/TweetsDetail.ts
