+++ src/frontend/worker/connect/delete.ts
@@ -2,7 +2,7 @@ import { State } from "frontend/globalState/types";
 import {
   TweetsDetailObj,
   TweetsDetail,
-} from "frontend/globalState/models/tweetsDetail";
+} from "frontend/globalState/models/TweetsDetail";
 import Immutable from "immutable";
 import { db } from "./db";
